Generate a summary of the available data.

Generate a summary of the available data.

fars_summarize_years(years)

Arguments

years

A vector of years for which data is desired.

Value

A single tibble summarizing the available data for each of the months in the provided vector of years

Errors

This function will throw an error if an unavailable or invalid year is encountered.

Examples

#Basic usage:
fars_summarize_years(years = c(2013, 2014, 2015))
#> # A tibble: 12 × 4
#>    MONTH `2013` `2014` `2015`
#>    <dbl>  <int>  <int>  <int>
#>  1     1   2230   2168   2368
#>  2     2   1952   1893   1968
#>  3     3   2356   2245   2385
#>  4     4   2300   2308   2430
#>  5     5   2532   2596   2847
#>  6     6   2692   2583   2765
#>  7     7   2660   2696   2998
#>  8     8   2899   2800   3016
#>  9     9   2741   2618   2865
#> 10    10   2768   2831   3019
#> 11    11   2615   2714   2724
#> 12    12   2457   2604   2781